JSUrl.jl Documentation

JSUrl — Module

Julia implementation of JSUrl. See the JSURL docs for more details on the format.

Methods

JSUrl.encode — Method

Encode a string type with JSUrl safe characters

JSUrl.parse — Method

Parse a JSUrl serialized string into a Julia object.

To deserialize your own type, look for a Dict with a single string key matching your typename. The value will be the output of repr(::YourType)

JSUrl.stringify — Method

Serialize a Julia object into a JSUrl string. This serialization is lossy as not all type information can be recovered.

To implement it for your own type, implement show(::YourType).

Index

JSUrl

JSUrl.encode

JSUrl.parse

JSUrl.stringify